+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ARENA-py notable changes. Started 2021-19-2.
 
+## [0.1.13] - 2021-19-2
+### Added
+- Support for scene landmarks with the `scene.add_landmark` method.
+
+### Changed
+- Bug fix with `scene.new_obj_callback` which didn't work well with code that updated `scene.users`.
+
 ## [0.1.12] - 2021-19-2
 ### Added
 - Support for user callbacks (`user_join_callback`, `user_left_callback`).